pyproject.toml
[tool.poetry]
name = "applybn"
version = "0.1.0"
description = ""
authors = ["jrzkaminski <jkaminski@niuitmo.ru>"]
readme = "README.md"
packages = [{include = "applybn"}]

[tool.poetry.dependencies]
python = "^3.11"

# miscellaneous
setuptools = "75.6.0"
rich = "^13.0.0"

# ML libraries
numpy = "^1.26.4"
pandas = "^2.0.0"
scipy = "^1.14.0"
scikit-learn = "^1.5.0"
imbalanced-learn = "^0.13.0"
torch = "^2.5.0"
statsmodels = "^0.14.0"
opencv-python = "4.11.0.86"

# causal inference libraries
econml = "0.15.1"
bamt = "^1.3.0"
jpype1 = "^1.5.2"

# ts anomaly detection
bayesian-optimization = "^2.0.3"


[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"
pytest-cov = "4.1.0"
pytest-mock = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
